+++ my-own-mentor-frontend/src/mentor/MentorSingleCard.js
@@ -1,55 +1,123 @@
-import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import Card from "@material-ui/core/Card";
-import CardActions from "@material-ui/core/CardActions";
-import CardContent from "@material-ui/core/CardContent";
-import Button from "@material-ui/core/Button";
-import Typography from "@material-ui/core/Typography";
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import clsx from 'clsx';
+import Card from '@material-ui/core/Card';
+import CardHeader from '@material-ui/core/CardHeader';
+import CardMedia from '@material-ui/core/CardMedia';
+import CardContent from '@material-ui/core/CardContent';
+import CardActions from '@material-ui/core/CardActions';
+import Collapse from '@material-ui/core/Collapse';
+import Avatar from '@material-ui/core/Avatar';
+import IconButton from '@material-ui/core/IconButton';
+import Typography from '@material-ui/core/Typography';
+import FavoriteIcon from '@material-ui/icons/Favorite';
+import ShareIcon from '@material-ui/icons/Share';
+import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
+import MoreVertIcon from '@material-ui/icons/MoreVert';
+import placeholder1 from "../ContentImages/jeremy-mcknight-ix9I7kfMpcc-unsplash.jpg"
 
-const useStyles = makeStyles({
+const useStyles = makeStyles((theme) => ({
     root: {
-        maxWidth: 275,
-        margin: "20px",
+        maxWidth: 375,
+        margin: "20px"
     },
-    bullet: {
-        display: "inline-block",
-        transform: "scale(0.8)"
+    media: {
+        height: 0,
+        paddingTop: '56.25%', // 16:9
     },
-    title: {
-        fontSize: 14
+    expand: {
+        transform: 'rotate(0deg)',
+        marginLeft: 'auto',
+        transition: theme.transitions.create('transform', {
+            duration: theme.transitions.duration.shortest,
+        }),
     },
-    pos: {
-        marginBottom: 12
-    }
-});
-export default function MentorSingleCard({firstName, lastName, university, description}) {
+    expandOpen: {
+        transform: 'rotate(180deg)',
+    },
+    avatar: {
+        backgroundColor: "secondary"
+    },
+}));
+
+export default function MentorSingleCard({ firstName, lastName, university, subject, description, pricing, email}) {
     const classes = useStyles();
-    const bull = <span className={classes.bullet}>•</span>;
+    const [expanded, setExpanded] = React.useState(false);
+
+    const handleExpandClick = () => {
+        setExpanded(!expanded);
+    };
     return (
         <Card className={classes.root}>
+            <CardHeader
+                avatar={
+                    <Avatar aria-label="recipe" className={classes.avatar}>
+                        R
+                    </Avatar>
+                }
+                action={
+                    <IconButton aria-label="settings">
+                        <MoreVertIcon />
+                    </IconButton>
+                }
+                title= {firstName}
+                subheader= {university}
+            />
+            <CardMedia
+                className={classes.media}
+                image={placeholder1}
+                title="Paella dish"
+            />
             <CardContent>
-                <Typography
-                    className={classes.title}
-                    color="textSecondary"
-                    gutterBottom
-                >
-                    {firstName} {lastName}
-                </Typography>
-                <Typography variant="h5" component="h2">
-                    {university}
-                </Typography>
-                <Typography className={classes.pos} color="textSecondary">
+                <Typography variant="body2" color="textSecondary" component="p">
                     {description}
                 </Typography>
-                <Typography variant="body2" component="p">
-                    well meaning and kindly.
-                    <br />
-                    {'"a benevolent smile"'}
-                </Typography>
             </CardContent>
-            <CardActions>
-                <Button size="small">Learn More</Button>
+            <CardActions disableSpacing>
+                <IconButton aria-label="add to favorites">
+                    <FavoriteIcon />
+                </IconButton>
+                <IconButton aria-label="share">
+                    <ShareIcon />
+                </IconButton>
+                <IconButton
+                    className={clsx(classes.expand, {
+                        [classes.expandOpen]: expanded,
+                    })}
+                    onClick={handleExpandClick}
+                    aria-expanded={expanded}
+                    aria-label="show more"
+                >
+                    <ExpandMoreIcon />
+                </IconButton>
             </CardActions>
+            <Collapse in={expanded} timeout="auto" unmountOnExit>
+                <CardContent>
+                    <Typography paragraph>Method:</Typography>
+                    <Typography paragraph>
+                        Heat 1/2 cup of the broth in a pot until simmering, add saffron and set aside for 10
+                        minutes.
+                    </Typography>
+                    <Typography paragraph>
+                        Heat oil in a (14- to 16-inch) paella pan or a large, deep skillet over medium-high
+                        heat. Add chicken, shrimp and chorizo, and cook, stirring occasionally until lightly
+                        browned, 6 to 8 minutes. Transfer shrimp to a large plate and set aside, leaving chicken
+                        and chorizo in the pan. Add pimentón, bay leaves, garlic, tomatoes, onion, salt and
+                        pepper, and cook, stirring often until thickened and fragrant, about 10 minutes. Add
+                        saffron broth and remaining 4 1/2 cups chicken broth; bring to a boil.
+                    </Typography>
+                    <Typography paragraph>
+                        Add rice and stir very gently to distribute. Top with artichokes and peppers, and cook
+                        without stirring, until most of the liquid is absorbed, 15 to 18 minutes. Reduce heat to
+                        medium-low, add reserved shrimp and mussels, tucking them down into the rice, and cook
+                        again without stirring, until mussels have opened and rice is just tender, 5 to 7
+                        minutes more. (Discard any mussels that don’t open.)
+                    </Typography>
+                    <Typography>
+                        Set aside off of the heat to let rest for 10 minutes, and then serve.
+                    </Typography>
+                </CardContent>
+            </Collapse>
         </Card>
     );
 }
